Add Spree::CustomerImage model
`Spree::CustomerImage` is the main model used to store customer image related
information.

Customers (`belongs_to :user`) can upload images (`has_one :image`) for a given
product (`belongs_to :product`).

Image and product are required, while the user is optional (not logged in
customers will also be able to upload images).

When a product is destroyed then all the customer uploaded images are destroyed
as well, while if a user goes away, the images are kept.

Customer images require admin moderation, so are created with boolean field
`approved` set to `NULL`, and they can be later approved or rejected. A scope
is provided for each of the 3 possible states.