Allow customers to upload images
The new section on the frontend allows customers to upload images for
a given product. The new form is accessible from the product page via
link injected with Deface. When customer is logged in then the email
field is automatically populated with their email.

Integration and controller specs are added.