Customer image model improvements
`email` field is now required. This field is autopopulated when the user is
already logged in, otherwise it must be provided.

The `image` association is automatically built when instantiating a new
`Spree::CustomerImage` record, if not present. The model also accepts nested
attributes for `image` in order to simplify the frontend form development.